Move gitlab.go to providers directory, update references
Renamed the 'internal/hooks/gitlab.go' file to 'internal/providers/gitlab.go' better representing its actual role within the application. All references to the old 'hooks' package within the 'internal/server/handlers.go' file have been updated to now refer to the new 'providers' package. This reorganization is more consistent with the overall application architecture and enhances code readability.

golang-ci clean up of unused code
